Exercise: Async and Sagas: Create reducer actions
First, make sure you have the necessary reducer actions that
you'll need when making your async requests.

For form submissions, you'll likely need actions such as:

1. Create/edit request - As far as the actual reducer function, you
likely won't need to do much here other than update the state of
the request itself.

2. Handle request result - Here is where you'll update the state with
the request response data, as well as update the state of the request.

3. Handle request error - If the request failed, you can store the
error response to display in the app, for example, as well as update
the status of the request.

4. Reset state - This is an option to to reset the data, such as
the request status.