Exercise: Async and Sagas: Create feature sagas
Next you'll create the saga functionality. You should have a separate
saga file for each async action that you'll be taking.

The saga should contain the following:
1. A watcher function that listens for a triggering action and
then calls a callback function
2. The callback function that includes the async functionality. The
callback function will generally also trigger a subsequent action
to the reducer based on the request response or error.

Once the saga is ready, you'll include it in the saga index so
that it will be included in the root saga.