Exercise: Async and Sagas BONUS: Pass params via React Navigation
To finish off the new edit functionality, we'll need to be able to
access an animal id once we've navigated to the Edit screen. These
changes break down passing the value as a route param which can then
be accessed by the screen logic and incorporated into our saga.

1. Update the feature reducer function in question to accept a payload.
(See `reducers.js`)

2. Update the feature saga function to take in the action as a param.
You can then pull the expected data from the `action.payload` and
use it as needed in your code. (See `sagas/edit.js`)

3. Next, we'll need to actually pass the value in the hook when
we dispatch the action. To gain access to the value in the first place,
we'll also have to pass the value as a param to the hook itself.
(See `hooks.js`)

4. When we call the hook, we'll have to now pass the value as the
new param value. For this particular example we'll expect to
receive the value from a prior screen that is passed via React
Navigation. To pull values from the route, we'll use the `useRoute`
hook and pull out the `params` from the route. (See `Edit/index.js`)

5. Finally, to pass the value through navigation, on the entry screen
we'll simply pass a params object to the `navigate` function that
includes our needed value. (See `row.js`)